Made gold into int. Fixed assempleFeatures. Fixed save and load tests

+++ src/main/java/Items/CoinBag.java
@@ -18,7 +18,7 @@ public class CoinBag {
         return this.goldCoin;
     }
 
-    public double getCoin(){
+    public int getCoin(){
         return this.goldCoin;
     }
